+++ internal/action/bufpane.go
@@ -527,7 +527,7 @@ func (h *BufPane) HandleEvent(event tcell.Event) {
 		if e.Buttons() != tcell.ButtonNone {
 			me := MouseEvent{
 				btn:   e.Buttons(),
-				mod:   metaToAlt(e.Modifiers()),
+				mod:   config.MetaToAlt(e.Modifiers()),
 				state: MousePress,
 			}
 			isDrag := len(h.mousePressed) > 0